requirements.txt
flask
flask_cors
flask_login
flask_wtf
flask_restful
flask_sqlalchemy
influxdb
influxalchemy
psutil
pyjwt
pyopenssl
PyYAML
git+https://github.com/theolind/pymysensors
